Greatly simplified the (un)installation process
The XVSR will now install all required assemblies and patches upon its first run. Core assemblies are copies to a subfolder or PrivateAssemblies in VS dir, and proper QTAgent*.exe.config is patched accordingly. Process is mostly automatic, but usually requires UAC check and a few clicks.

Also, a simple command was added to the SolutionExplorer's Project context menu, and it allows to quickly mark/unmark the project as a 'test project'.

Also a 'kill QTAgent' tool was added to the main 'Test' menu of VS IDE.