Autoinstaller now patches all QTAgent.exe.configs
Aside from the tiny workflow change, a number of changes had to be made to the MainWindow, as there are now many state entries for the config files. Those are now presented in the same manner as assembly status, with small change - the rows describing configuration changes are considered 'child' while assembly rows - 'parent'. Any change to assembly state resets all configuration rows' actions to their defaults with respect to new assembly states.